Release 1.1.0: bug-hunt fixes
Found by fuzzing, ThreadSanitizer stress and review:
- Tar reader: sizes from headers are checked against the archive size,
  and pax/long-name records capped at 1 MiB; a corrupt record claiming
  exabytes made the GUI try to allocate it and crash. Truncated archives
  keep the members they fully contain.
- Tar reader: members with a pax size override (files over 8 GiB) are
  skipped by their real size instead of the ustar field, which landed
  mid-file and broke the rest of the archive.
- abp gui handles SIGINT/SIGTERM: the job is cancelled, the server stops
  and the explorer cache in /tmp is removed.
- uninstall.sh only removes a binary that reports itself as abp.
- Gallery video tiles load their source only when scrolled into view.
- strerror() replaced with the thread-safe generic_category().message().

Release: version 1.1.0, CHANGELOG section dated, stray duplicate
"Added" heading folded in, version references in the docs updated.

+++ src/util/Json.cpp
@@ -319,7 +319,7 @@ void JsonValue::dumpTo(std::string& out, int indent, int depth) const {
     auto writeIndent = [&](int level) {
         if (indent > 0) {
             out.push_back('\n');
-            out.append(static_cast<size_t>(indent * level), ' ');
+            out.append(static_cast<size_t>(indent) * static_cast<size_t>(level), ' ');
         }
     };
